requirements.txt
absl-py==2.3.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==0.9.9
annotated-types==0.5.0
anyio==4.9.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.1.0
bcrypt==3.2.2
beautifulsoup4==4.13.4
betterproto==1.2.5
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.19.1
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.4.2
click==8.3.0
contourpy==1.2.1
cryptography==42.0.7
curl-cffi==0.5.7
cycler==0.12.1
decorator==5.2.1
deepface==0.0.93
Deprecated==1.2.13
deprecation==2.1.0
discordwebhook==1.0.3
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.2
flask-cors==6.0.0
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Limiter==3.3.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flatbuffers==25.2.10
fonttools==4.53.0
frozenlist==1.4.0
fsspec==2025.9.0
g4f==0.2.8.2
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
gitignore_parser==0.1.6
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.25.0
google-api-python-client==2.137.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.21.1
google-generativeai==0.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==1.3.1
greenlet==3.1.1
grpcio==1.64.1
grpcio-status==1.62.2
grpclib==0.4.5
gunicorn==20.1.0
h11==0.16.0
h2==4.1.0
h5py==3.13.0
hf-xet==1.2.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.36.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.10
importlib-resources==5.12.0
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.6
joblib==1.5.1
jsonpickle==3.0.2
jwt==1.3.1
keras==3.10.0
kiwisolver==1.4.5
lazy==1.3
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
limits==3.4.0
llvmlite==0.45.1
lz4==4.4.4
Mako==1.0.7
Markdown==3.8
markdown-it-py==2.2.0
MarkupSafe==3.0.3
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mongoengine==0.27.0
more-itertools==10.8.0
motor==3.4.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.18.6
mtcnn==1.0.0
multidict==6.0.4
Naked==0.1.32
namex==0.1.0
networkx==3.3
numba==0.62.1
numpy==1.26.4
oauth2==1.9.0.post1
oauthlib==2.0.7
ollama==0.2.1
openai==1.16.2
openai-whisper==20250625
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
ordered-set==4.1.0
packaging==23.1
pandas==2.2.3
pillow==10.3.0
platformdirs==4.5.0
poe-api==0.5.2
pooch==1.8.2
postgrest==0.11.0
praat-parselmouth==0.4.6
priority==2.0.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.18.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.15.1
PyJWT==2.8.0
pymongo==4.6.2
Pympler==1.0.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.7.2
python-dotenv==1.1.0
python-editor==1.0.3
python-jose==3.3.0
python-socks==2.2.0
pytz==2025.2
PyYAML==6.0.1
pyzmq==25.1.1
Quart==0.19.6
quart-bcrypt==0.0.7
quart-cors==0.7.0
quickjs==1.19.2
realtime==1.0.6
regex==2025.10.23
requests==2.32.4
requests-oauthlib==0.8.0
retina-face==0.0.17
rich==13.3.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==69.2.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
storage3==0.6.1
StrEnum==0.4.15
stringcase==1.2.0
supabase==2.3.4
supafunc==0.3.3
sympy==1.12.1
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==3.1.0
tf_keras==2.16.0
thefuzz==0.19.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tls-client==0.2.1
tokenizers==0.22.1
torch==2.2.2
tqdm==4.66.2
transformers==4.57.1
twelvelabs==0.4.10
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
URLObject==2.4.3
websocket-client==1.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
wincertstore==0.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==2.1
yarl==1.9.2
